+++ strategies/rsi_strategy.py
@@ -49,7 +49,8 @@ class RSIStrategy(BaseStrategy):
         # =========================
         # PERSIST POSITION (FIXED)
         # =========================
+        df["signal"] = df["signal"].astype('Int64')  # Use nullable integer
         df["signal"] = df["signal"].replace(0, pd.NA).ffill()
-        df["signal"] = df["signal"].fillna(0)
+        df["signal"] = df["signal"].fillna(0).astype(int)
 
         return df
